+++ plugins/size/src/formatUtils.ts
@@ -1,10 +1,12 @@
 import {
-    SizeResult
+    SizeResult,
+    Export
 } from "./interfaces"
 import fileSize from 'file-size';
 import colorette from 'colorette';
 
 const FAILURE_THRESHOLD = 5;
+const RUNTIME_SIZE = 537;
 
 /** Format the size to a human readable format. */
 const format = (size: number) => {
@@ -95,4 +97,54 @@ const defaultTotals = (results: SizeResult[], css?: boolean) => {
     ];
 };
 
-export { formatLine, defaultTotals }
+/** Get the size of a chunk */
+const sumChunk = (exported: Export[], chunkName: string) =>
+  exported
+    .filter(m => m.chunkNames.includes(chunkName))
+    .reduce((acc, i) => acc + i.size, 0);
+
+/** Format a export lines of the output table. */
+const formatExports = (
+    { master, pr }: SizeResult,
+    css?: boolean,
+    preName = ''
+  ) => {
+    const masterExports = master.exported || [];
+    const masterCSS = sumChunk(masterExports, 'css');
+    const prExports = pr.exported || [];
+    const prCSS = sumChunk(prExports, 'css');
+    const lines: (string | number)[][] = [];
+  
+    const chunks = prExports.reduce((acc, i) => {
+      i.chunkNames.forEach(name => acc.add(name));
+      return acc;
+    }, new Set<string>());
+  
+    chunks.forEach(chunkName => {
+      if (chunkName === 'css') {
+        return;
+      }
+  
+      const masterChunks = sumChunk(masterExports, chunkName);
+      const prChunks = sumChunk(prExports, chunkName);
+  
+      lines.push([
+        `${preName}${chunkName}`,
+        ...formatLine(
+          {
+            master: {
+              css: masterCSS,
+              js: masterChunks ? masterChunks - RUNTIME_SIZE : 0
+            },
+            pr: { css: prCSS, js: prChunks - RUNTIME_SIZE },
+            percent: 0
+          },
+          css
+        )
+      ]);
+    });
+  
+    return lines;
+  };
+
+export { formatLine, defaultTotals, formatExports }
